+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - 
 ### Changed
 - changed table color in the Excel exporters
-- 
+- refactored FDD Fault Alarm database, api and web
 ### Fixed
 - updated combined equipment energy category excel exporter to show parameters data even if there is no energy data 
 ### Removed